Moved back endpoints/test_utils.go to endpoints/integration_test.go
